try to solve cycle import but fail
