+++ pkg/planner/core/operator/physicalop/plan_clone_generated.go
@@ -21,6 +21,7 @@ import (
 	"github.com/pingcap/tidb/pkg/planner/core/base"
 	"github.com/pingcap/tidb/pkg/planner/util"
 	"github.com/pingcap/tidb/pkg/planner/util/utilfuncp"
+	sliceutil "github.com/pingcap/tidb/pkg/util/slice"
 )
 
 // CloneForPlanCache implements the base.Plan interface.
@@ -107,7 +108,7 @@ func (op *PhysicalTableScan) CloneForPlanCache(newCtx base.PlanContext) (base.Pl
 	if op.HandleCols != nil {
 		cloned.HandleCols = op.HandleCols.Clone()
 	}
-	cloned.ByItems = util.CloneByItemss(op.ByItems)
+	cloned.ByItems = sliceutil.SliceDeepClone(op.ByItems)
 	cloned.PlanPartInfo = op.PlanPartInfo.CloneForPlanCache()
 	if op.SampleInfo != nil {
 		return nil, false
@@ -139,7 +140,7 @@ func (op *PhysicalIndexScan) CloneForPlanCache(newCtx base.PlanContext) (base.Pl
 	if op.GenExprs != nil {
 		return nil, false
 	}
-	cloned.ByItems = util.CloneByItemss(op.ByItems)
+	cloned.ByItems = sliceutil.SliceDeepClone(op.ByItems)
 	if op.PkIsHandleCol != nil {
 		if op.PkIsHandleCol.SafeToShareAcrossSession() {
 			cloned.PkIsHandleCol = op.PkIsHandleCol
@@ -187,8 +188,8 @@ func (op *PhysicalTopN) CloneForPlanCache(newCtx base.PlanContext) (base.Plan, b
 		return nil, false
 	}
 	cloned.PhysicalSchemaProducer = *basePlan
-	cloned.ByItems = util.CloneByItemss(op.ByItems)
-	cloned.PartitionBy = util.CloneSortItems(op.PartitionBy)
+	cloned.ByItems = sliceutil.SliceDeepClone(op.ByItems)
+	cloned.PartitionBy = sliceutil.SliceDeepClone(op.PartitionBy)
 	return cloned, true
 }
 
@@ -201,7 +202,7 @@ func (op *PhysicalLimit) CloneForPlanCache(newCtx base.PlanContext) (base.Plan, 
 		return nil, false
 	}
 	cloned.PhysicalSchemaProducer = *basePlan
-	cloned.PartitionBy = util.CloneSortItems(op.PartitionBy)
+	cloned.PartitionBy = sliceutil.SliceDeepClone(op.PartitionBy)
 	return cloned, true
 }
 
@@ -354,7 +355,7 @@ func (op *PhysicalIndexMergeReader) CloneForPlanCache(newCtx base.PlanContext) (
 	}
 	cloned.PhysicalSchemaProducer = *basePlan
 	cloned.PushedLimit = op.PushedLimit.Clone()
-	cloned.ByItems = util.CloneByItemss(op.ByItems)
+	cloned.ByItems = sliceutil.SliceDeepClone(op.ByItems)
 	PartialPlansRaw, ok := ClonePhysicalPlansForPlanCache(newCtx, op.PartialPlansRaw)
 	if !ok {
 		return nil, false
