+++ pkg/planner/property/physical_property.go
@@ -87,6 +87,21 @@ func (s SortItem) MemoryUsage() (sum int64) {
 	return
 }
 
+// ExplainPartitionBy produce text for p.PartitionBy. Common for window functions and TopN.
+func ExplainPartitionBy(ctx expression.EvalContext, buffer *bytes.Buffer,
+	partitionBy []SortItem, normalized bool) *bytes.Buffer {
+	if len(partitionBy) > 0 {
+		buffer.WriteString("partition by ")
+		for i, item := range partitionBy {
+			fmt.Fprintf(buffer, "%s", item.Col.ColumnExplainInfo(ctx, normalized))
+			if i+1 < len(partitionBy) {
+				buffer.WriteString(", ")
+			}
+		}
+	}
+	return buffer
+}
+
 // MPPPartitionType is the way to partition during mpp data exchanging.
 type MPPPartitionType int
