+++ pkg/planner/util/misc.go
@@ -18,7 +18,6 @@ import (
 	"encoding/binary"
 	"fmt"
 	"iter"
-	"math"
 	"reflect"
 	"slices"
 	"time"
@@ -27,30 +26,14 @@ import (
 	"github.com/pingcap/tidb/pkg/expression"
 	"github.com/pingcap/tidb/pkg/kv"
 	"github.com/pingcap/tidb/pkg/meta/metadef"
-	"github.com/pingcap/tidb/pkg/meta/model"
 	"github.com/pingcap/tidb/pkg/parser/ast"
 	"github.com/pingcap/tidb/pkg/planner/core/base"
-	"github.com/pingcap/tidb/pkg/planner/property"
 	"github.com/pingcap/tidb/pkg/sessionctx/vardef"
 	"github.com/pingcap/tidb/pkg/types"
 	"github.com/pingcap/tidb/pkg/util/dbterror/plannererrors"
 	h "github.com/pingcap/tidb/pkg/util/hint"
 	"github.com/pingcap/tidb/pkg/util/intest"
-	"github.com/pingcap/tidb/pkg/util/ranger"
 )
-
-// SliceDeepClone uses Clone() to clone a slice.
-// The elements in the slice must implement func (T) Clone() T.
-func SliceDeepClone[T interface{ Clone() T }](s []T) []T {
-	if s == nil {
-		return nil
-	}
-	cloned := make([]T, 0, len(s))
-	for _, item := range s {
-		cloned = append(cloned, item.Clone())
-	}
-	return cloned
-}
 
 // SliceRecursiveFlattenIter returns an iterator (iter.Seq2) that recursively iterates over all elements of an
 // any-dimensional slice of any type.
@@ -119,16 +102,6 @@ func CloneFieldNames(names []*types.FieldName) []*types.FieldName {
 	return cloned
 }
 
-// CloneCIStrs uses ast.CIStr.Clone to clone a slice of ast.CIStr.
-func CloneCIStrs(strs []ast.CIStr) []ast.CIStr {
-	if strs == nil {
-		return nil
-	}
-	cloned := make([]ast.CIStr, 0, len(strs))
-	cloned = append(cloned, strs...)
-	return cloned
-}
-
 // CloneExprs uses Expression.Clone to clone a slice of Expression.
 func CloneExprs(exprs []expression.Expression) []expression.Expression {
 	if exprs == nil {
@@ -139,11 +112,6 @@ func CloneExprs(exprs []expression.Expression) []expression.Expression {
 		cloned = append(cloned, e.Clone())
 	}
 	return cloned
-}
-
-// CloneExpressions uses CloneExprs to clone a slice of expression.Expression.
-func CloneExpressions(exprs []expression.Expression) []expression.Expression {
-	return CloneExprs(exprs)
 }
 
 // CloneAssignments uses (*Assignment).Clone to clone a slice of *Assignment.
@@ -186,18 +154,6 @@ func CloneCols(cols []*expression.Column) []*expression.Column {
 	return cloned
 }
 
-// CloneConstants uses (*Constant).Clone to clone a slice of *Constant.
-func CloneConstants(constants []*expression.Constant) []*expression.Constant {
-	if constants == nil {
-		return nil
-	}
-	cloned := make([]*expression.Constant, 0, len(constants))
-	for _, c := range constants {
-		cloned = append(cloned, c.Clone().(*expression.Constant))
-	}
-	return cloned
-}
-
 // CloneDatums uses Datum.Clone to clone a slice of Datum.
 func CloneDatums(datums []types.Datum) []types.Datum {
 	if datums == nil {
@@ -218,54 +174,6 @@ func CloneDatum2D(datums [][]types.Datum) [][]types.Datum {
 	cloned := make([][]types.Datum, 0, len(datums))
 	for _, d := range datums {
 		cloned = append(cloned, CloneDatums(d))
-	}
-	return cloned
-}
-
-// CloneColInfos uses (*ColumnInfo).Clone to clone a slice of *ColumnInfo.
-func CloneColInfos(cols []*model.ColumnInfo) []*model.ColumnInfo {
-	if cols == nil {
-		return nil
-	}
-	cloned := make([]*model.ColumnInfo, 0, len(cols))
-	for _, c := range cols {
-		cloned = append(cloned, c.Clone())
-	}
-	return cloned
-}
-
-// CloneRanges uses (*Range).Clone to clone a slice of *Range.
-func CloneRanges(ranges []*ranger.Range) []*ranger.Range {
-	if ranges == nil {
-		return nil
-	}
-	cloned := make([]*ranger.Range, 0, len(ranges))
-	for _, r := range ranges {
-		cloned = append(cloned, r.Clone())
-	}
-	return cloned
-}
-
-// CloneByItemss uses (*ByItems).Clone to clone a slice of *ByItems.
-func CloneByItemss(byItems []*ByItems) []*ByItems {
-	if byItems == nil {
-		return nil
-	}
-	cloned := make([]*ByItems, 0, len(byItems))
-	for _, item := range byItems {
-		cloned = append(cloned, item.Clone())
-	}
-	return cloned
-}
-
-// CloneSortItems uses SortItem.Clone to clone a slice of SortItem.
-func CloneSortItems(items []property.SortItem) []property.SortItem {
-	if items == nil {
-		return nil
-	}
-	cloned := make([]property.SortItem, 0, len(items))
-	for _, item := range items {
-		cloned = append(cloned, item.Clone())
 	}
 	return cloned
 }
@@ -326,20 +234,6 @@ func GetMaxSortPrefix(sortCols, allCols []*expression.Column) []int {
 		sortColOffsets = append(sortColOffsets, offset)
 	}
 	return sortColOffsets
-}
-
-// DeriveLimitStats derives the stats of the top-n plan.
-func DeriveLimitStats(childProfile *property.StatsInfo, limitCount float64) *property.StatsInfo {
-	stats := &property.StatsInfo{
-		RowCount: math.Min(limitCount, childProfile.RowCount),
-		ColNDVs:  make(map[int64]float64, len(childProfile.ColNDVs)),
-		// limit operation does not change the histogram (kind of sample).
-		HistColl: childProfile.HistColl,
-	}
-	for id, c := range childProfile.ColNDVs {
-		stats.ColNDVs[id] = math.Min(c, stats.RowCount)
-	}
-	return stats
 }
 
 // ExtractTableAlias returns table alias of the base.LogicalPlan's columns.
