+++ pkg/planner/util/path.go
@@ -26,6 +26,7 @@ import (
 	"github.com/pingcap/tidb/pkg/types"
 	"github.com/pingcap/tidb/pkg/util/collate"
 	"github.com/pingcap/tidb/pkg/util/ranger"
+	sliceutil "github.com/pingcap/tidb/pkg/util/slice"
 )
 
 // AccessPath indicates the way we access a table: by using single index, or by using multiple indexes,
@@ -143,7 +144,7 @@ func (path *AccessPath) Clone() *AccessPath {
 		IdxCols:                      CloneCols(path.IdxCols),
 		IdxColLens:                   slices.Clone(path.IdxColLens),
 		ConstCols:                    slices.Clone(path.ConstCols),
-		Ranges:                       CloneRanges(path.Ranges),
+		Ranges:                       sliceutil.SliceDeepClone(path.Ranges),
 		CountAfterAccess:             path.CountAfterAccess,
 		MinCountAfterAccess:          path.MinCountAfterAccess,
 		MaxCountAfterAccess:          path.MaxCountAfterAccess,
@@ -170,12 +171,12 @@ func (path *AccessPath) Clone() *AccessPath {
 	if path.IndexMergeORSourceFilter != nil {
 		ret.IndexMergeORSourceFilter = path.IndexMergeORSourceFilter.Clone()
 	}
-	ret.PartialIndexPaths = SliceDeepClone(path.PartialIndexPaths)
+	ret.PartialIndexPaths = sliceutil.SliceDeepClone(path.PartialIndexPaths)
 	ret.PartialAlternativeIndexPaths = make([][][]*AccessPath, 0, len(path.PartialAlternativeIndexPaths))
 	for _, oneORBranch := range path.PartialAlternativeIndexPaths {
 		clonedORBranch := make([][]*AccessPath, 0, len(oneORBranch))
 		for _, oneAlternative := range oneORBranch {
-			clonedOneAlternative := SliceDeepClone(oneAlternative)
+			clonedOneAlternative := sliceutil.SliceDeepClone(oneAlternative)
 			clonedORBranch = append(clonedORBranch, clonedOneAlternative)
 		}
 		ret.PartialAlternativeIndexPaths = append(ret.PartialAlternativeIndexPaths, clonedORBranch)
